refactor: move `Leaderboard` to separate module
Moves the `Leaderboard` struct and impl into a dedicated mod.
