+++ src/leaderboard/mod.rs
@@ -6,8 +6,10 @@ pub use messages::MomentoRequest;
 mod config;
 mod leaderboard_client;
 mod leaderboard_client_builder;
+mod leaderboard_resource;
 
 pub use config::configuration::Configuration;
 pub use config::configurations;
 
-pub use leaderboard_client::{Leaderboard, LeaderboardClient};
+pub use leaderboard_client::LeaderboardClient;
+pub use leaderboard_resource::Leaderboard;
